refactor(graph): extract graph_parallel transformer + doc/wrap CI fixes
Address the three Cursor-agent CI gates on PR #756:

- Feature factorization: move the graph_parallel detection/validation/lowering
  out of ast_transformer.py into ast_transformers/graph_parallel_transformer.py
  (GraphParallelTransformer), mirroring checkpoint_transformer.py. build_With now
  calls the new class directly; function_def_transformer comments updated to point
  at it.
- Line wrapping: re-wrap the graph_parallel_context docstring to 120c and drop a
  dangling reference to an internal design doc.
- Doc quality (graph.md): reword pre-existing graph_do_while content the gate
  flagged -- replace "lowering" jargon, drop the internal "genesis-world" /
  hypothetical qd.graph_range_for / MAX_ITER design walk-through phrasing.
